api: Move SiaeSerializer
